80Authorization with Gates

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -10,6 +10,11 @@ use Illuminate\Http\Request;
 
 class AttendeeController extends Controller
 {
+    public function __construct()
+    {
+        //* 除了index, show其他 Event的操作都需要先經驗證
+        $this->middleware('auth:sanctum')->except('index', 'show', 'update');
+    }
     /**
      * Display a listing of the resource.
      */
@@ -51,8 +56,11 @@ class AttendeeController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $event, Attendee $attendee)
+    public function destroy(Event $event, Attendee $attendee)
     {
+        //* 事件的所有者或參與者才能改
+        $this->authorize('delete-attendee', [$event, $attendee]);
+
         $attendee->delete();
 
         return response(status: 204);
